refactor: add converters.h
